google-chat/requirements.txt
httplib2
pyyaml